Functions/Top-hoteles/requirements.txt
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
cloudevents==1.10.1
db-dtypes==1.1.1
deprecation==2.1.0
google-api-core==2.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-cloud-pubsub==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
idna==3.4
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.2
pandas-gbq==0.19.2
proto-plus==1.22.3
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydata-google-auth==1.8.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
yelpapi==2.5.1